第2-4节：策略模式优惠折扣计算

+++ group-buy-market-zsy-infrastructure/src/main/java/cn/bugstack/infrastructure/dao/po/GroupBuyDiscount.java
@@ -36,7 +36,7 @@ public class GroupBuyDiscount {
     /**
      * 折扣类型（0:base、1:tag）
      */
-    private Byte discountType;
+    private Integer discountType;
 
     /**
      * 营销优惠计划（ZJ:直减、MJ:满减、N元购）
